split map commands into two classes
